last fix

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -8,7 +8,7 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t count;
+	long unsigned int count;
 	const list_t *ptr;
 	char *temp;
